make client operations synchronous (#26)
The goal of this commit is to to make the "default" ModuleClient and DeviceClient operations synchronous.

Later we will add async counterparts for Python versions that support the async/await syntax, and this design of this change takes future async/await functionality into consideration. Specifically, the choice of using a node-like callback parameter on transport operations. That callback is called by the transport when a specific operation (connect, disconnect, and send_event) is complete. In the current (sync) implementation, that callback is just used to set an event that the main thread is waiting for. In future (async/await) implementations, that callback can be used to complete a Future which is being awaited by the client code.

In order to make this work consistently, and in order to fix some problems that Pierre saw when trying to call send_event without connecting, I decided to simplify the state machine. Since we are looking forward to async/await and want to support overlapped sends, I decided to remove the "sending" state. This was made possible by the addition of the callback parameter, so each send_event operation can have its own callback, and there is no reason to leave the state machine in a "sending" state while we want for the acknowledgement from the service. Now we have 4 states: disconnected, connecting, connected, and disconnecting. For operations (like send_event) that happen while disconnected, we have an event queue that doesn't get processed until we're in the connected state.

I also realized that the Locked state machine does not behave the way I thought when dealing with overlapped transitions. I thought that transitions would be queued if another transition was in progress, but it turns out that transitions are only blocked while another transition is in progress. This is the root cause of Pierre's problem with send_event. Part of the simplification allowed me to change the machine into a queued state machine. This means that transitions will be queued if another transition is in progress. This also means that we don't have to worry about what we do in before, after, on_enter, and on_exit functions because any transitions that are triggered inside those functions are queued rather than executed immediately.

I'm happy that the unit tests didn't have to change much. This makes me feel confident about these changes.