POC for adding _async wrappers for InternalClient functions (#28)
This is one way we could create async wrappers for our code. I like this because:

All transport calls run in a background thread, so we never block the main event loop.
transport callbacks signal the awaiting thread when operations are done by using a Future object. This way the main event loop isn't blocked waiting operations to complete, and we don't need to make a thread that waits for our operation to complete.
async methods are added to InternalClient with _async suffixes. We could also make InternalClientAsync to hold the async functions, but that seems like it could explode as we add DeviceClientAsync and ModuleClientAsync, and I'm sure we would need to add more.
I've tested this with sending 1000 events in parallel and it seems to do as well as we want. I've also tested that we can still use the client object with Python 2.7 because the code that adds the _async implementations is protected by try/except.