add sas token renewal logic for shared key auth and iotedge auth (#33)
This adds code for the shared key and iotedge auth providers to automatically re-generate their sas tokens based on expiry timers.  In order to do this, several things needed to happen:
1) The MQTT provider needed to have a reconnect() function.  This is how the upper layers send a new SAS token down to PAHO.
2) The MQTT transport needed to have a callback function where the auth provider could notify the
transport that a new SAS token is available.  Since this operation requires the transport to reconnect, I needed to add a new trigger to the state machine.  There is currently no need to add a "reconnecting" state since "reconnecting" would behave exactly like "connecting".
3) A timer needed to be added to the auth providers so they could automatically renew themselves, and a callback needed to be defined so they could notify the transport when this renewal happens.  
4) A "disconnect" function needed to be added to the auth providers so they can cancel their update timers.  This is necessary because otherwise, a timer would prevent the auth object from being garbage collected at the right time.
5) The transport disconnect function needs to call disconnect on the auth provider so the application can shut down after disconnect is called.  Tests also need to call disconnect on the auth provider for the same reason.