bug(azure-iot-device): SAS Token renewal in HTTP Pipeline (#437)
There was an issue in the SAS Token renewal where the HTTP Pipeline would fail to connect due to an expired SAS token after an hour. 

We root caused it to be because we reuse the auth provider for both the HTTP and MQTT pipeline. The MQTT pipeline had overwritten the SAS Renewal handler. Now, instead of a single handler there is a list of handlers that is called each time the SAS token renews.