(fix): Add watchdog timer for MQTT connect and reconnect operations (#516)
There are conditions where connect and reconnect operations can timeout and we have no control.  These are very uncommon, but reproducable with the right stress scripts.  

# How did I fix this?

To work around these conditions, I've added a watchdog timer to connect and reconnect operations.  If these operations don't complete within some interval (I picked 10 seconds), the watchdog cancels the operation and disconnects the transport.  This gives the transport a "clean slate" and the pipeline retry/reconnect logic is able to recover.

# Problem 1: Paho conect never returns

In some cases, the call into paho's mqtt_client.connect() never returns or takes a really long time to return.  (this has been seen in the field, but I haven't been able to reproduce this).  This is likely a bug in TLS  negotiation or the network stack.

We have no way to fix this.

# Problem 2: Paho connect succeeds, but CONACK never comes back.

In other cases, connect returns success, but the CONACK never returns and we never get an on_mqtt_disconneted event.  I can reproduce this under stress situations, and it has to do with the fact that paho's loop isn't really entirely 100% thread safe, especially if you're intermixing bad network conditions with sas token renewal and DeviceClient opreations.

We could fix this, but it would involve some combination of replacing Paho's thread with our own and submitting PRs to Paho.  Both of these are really expensive.

# Why add the code here?

I did this in the MQTT stage because I wanted to keep the recovery code as close to the transport as possible.  

I could have used TimeoutStage to detect the timeout, but I would have needed to add a recovery stage to disconnect when timeouts occur.  This is working around behaviors of Paho, so it seemed silly to put into a generic stage.

I could have created a new stage to add watchdog functionality to these 2 ops, but this seemed like overkill.  Stages aren't complex, the the logic of ordering stages is complex because of implied inter-stage dependencies.